Added post method for sales invoices. Change type from int to long for almost all of IDs.

+++ Models/VatRate.cs
@@ -2,7 +2,7 @@
 {
     public class VatRate
     {
-        public int RateId { get; set; }
+        public long RateId { get; set; }
         public string Rate { get; set; }
         public string Symbol { get; set; }
         public string CountryCode { get; set; }
